Дополнительные
"между методами нужно оставлять одну пустую строку (enter), а то получается каша" (добавлена)

"не хватает в файле табуляции" (форматирование кода по умолчанию в VS)

"Комментарии если и нужны, то только на ангийском языке. В данном случае комментарий бесполезен." (убраны комментарии)

"Этот парсинг должен быть частью Arithm (имя неудачное)" (Переименован на ArithmParse. Название класса выбрана из-за того, что там выполняется парсинг и арифметические операции. Возможно, парсинг следует вынести в отдельный класс)
"дублирование строк, лучше завести константы" (исправлено, если правильно понял правку)
